backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
dashscope==1.14.1
google-generativeai==0.3.2
openai==1.3.7
volcengine-python-sdk==1.0.100
websockets==12.0 